homework2/requirements.txt
asgiref==3.7.2
Django==4.2.11
django-bootstrap5==25.1
djangorestframework==3.15.2
gyp==0.1
pytz==2024.1
setuptools==68.1.2
six==1.16.0
sqlparse==0.4.4
supervisor==4.2.5
wheel==0.42.0
